+++ CHANGELOG.rst
@@ -1,5 +1,6 @@
 1.1.1 (2015-XX-XX)
 ------------------
+- Fix unicode in model docstrings
 - Require swagger-spec-validator 1.0.12 to pick up bug fixes
 
 1.1.0 (2015-06-25)